+++ sys/dev/wscons/mra.c
@@ -55,7 +55,7 @@ mra_Y_AX1_BX2_C(const int *y, int ys,
 //	int64_t SYY;
 	int64_t S1Y, S2Y;
 	int64_t A, B, C, M;
-#define AA(p, s, i)	(*((const long *)(((const char *)(p)) + (s) * (i))))
+#define AA(p, s, i)	(*((const int *)(((const char *)(p)) + (s) * (i))))
 #define X1(i)		AA(x1, x1s, i)
 #define X2(i)		AA(x2, x2s, i)
 #define Y(i)		AA(y, ys, i)
